+++ src/app/locales/en/components/sidebar-spaces.js
@@ -8,6 +8,7 @@ export default {
     index: 'Overview',
     data: 'Data',
     transfers: 'Transfers',
+    shares: 'Shares',
     providers: 'Providers',
     members: 'Members',
     membership: 'Membership',