+++ src/app/utils/generate-development-model.js
@@ -12,7 +12,7 @@ import { camelize } from '@ember/string';
 import _ from 'lodash';
 import { A } from '@ember/array';
 import { Promise, resolve, all as allFulfilled, hash as hashFulfilled } from 'rsvp';
-import { get, set } from '@ember/object';
+import { get, set, setProperties } from '@ember/object';
 import groupPrivilegesFlags from 'onedata-gui-websocket-client/utils/group-privileges-flags';
 import spacePrivilegesFlags from 'onedata-gui-websocket-client/utils/space-privileges-flags';
 import harvesterPrivilegesFlags from 'onedata-gui-websocket-client/utils/harvester-privileges-flags';
@@ -20,7 +20,10 @@ import { inviteTokenSubtypeToTargetModelMapping } from 'onezone-gui/models/token
 import parseGri from 'onedata-gui-websocket-client/utils/parse-gri';
 import gri from 'onedata-gui-websocket-client/utils/gri';
 import moment from 'moment';
-import { generateSpaceEntityId } from 'onedata-gui-websocket-client/utils/development-model-common';
+import {
+  generateSpaceEntityId,
+  generateShareEntityId,
+} from 'onedata-gui-websocket-client/utils/development-model-common';
 
 const USER_ID = 'stub_user_id';
 const USERNAME = 'Stub User';
@@ -39,6 +42,7 @@ const PROVIDER_NAMES = ['Cracow', 'Paris', 'Lisbon'].concat(
 const types = [
   'space',
   'group',
+  'share',
   'provider',
   'linkedAccount',
   'cluster',
@@ -134,16 +138,47 @@ export default function generateDevelopmentModel(store) {
       const providers = listRecords.provider.get('list');
       const spaces = listRecords.space.get('list');
       const clusters = listRecords.cluster.get('list');
-      return allFulfilled([providers, spaces, clusters])
-        .then(([providerList, spaceList, clusterList]) =>
+      const shares = listRecords.share.get('list');
+      return allFulfilled([providers, spaces, clusters, shares])
+        .then(([providerList, spaceList, clusterList, shareList]) =>
           allFulfilled(spaceList.map(space => {
             space.set('supportSizes', _.zipObject(
               get(providers, 'content').mapBy('entityId'),
               _.fill(Array(NUMBER_OF_PROVIDERS), perProviderSize)
             ));
-            return createListRecord(store, 'provider', providerList).then(lr => {
-              space.set('providerList', lr);
-              return space.save();
+            return allFulfilled([
+              createListRecord(store, 'provider', providerList),
+              createListRecord(store, 'share', shareList),
+            ]).then(([providerLr, shareLr]) => {
+              const generalGriData = {
+                entityType: 'share',
+                entityId: generateShareEntityId(get(space, 'entityId')),
+                aspect: 'instance',
+              };
+              const privateGri = gri(Object.assign({
+                scope: 'private',
+              }, generalGriData));
+              const generalData = {
+                name: `Share for ${get(space, 'name')}`,
+                chosenProviderId: getProviderId(0),
+                chosenProviderVersion: '20.02.0-beta1',
+                fileType: 'dir',
+              };
+              return store.createRecord('share', Object.assign({
+                  id: privateGri,
+                }, generalData))
+                .save()
+                .then(share => {
+                  get(shareLr, 'list').pushObject(share);
+                  return shareLr.save();
+                })
+                .then(() => {
+                  setProperties(space, {
+                    providerList: providerLr,
+                    shareList: shareLr,
+                  });
+                  return space.save();
+                });
             });
           }))
           .then(() => allFulfilled(clusterList.map(cluster => {
@@ -296,6 +331,8 @@ function createEntityRecords(store, type, names, additionalInfo) {
       return createProvidersRecords(store, additionalInfo);
     case 'space':
       return createSpacesRecords(store, additionalInfo);
+    case 'share':
+      return createSharesRecords(store, additionalInfo);
     case 'token':
       return createTokensRecords(store, additionalInfo);
     case 'group':
@@ -395,6 +432,11 @@ function createSpacesRecords(store) {
       },
     }).save();
   }));
+}
+
+// currently shares records are created separately for each space
+function createSharesRecords() {
+  return resolve([]);
 }
 
 function createTokensRecords(store) {